refactor: consolidate research module — fewer files, same API
Move RunResult, RunStatus, ResultParseError to factory/models.py (alongside
ResearchTarget). Merge parser and store functions into runner.py to reduce
module count from 4 to 1.
